+++ src/custom-data-generator/cem-utilities.ts
@@ -1,52 +1,208 @@
-import { Declaration, Reference } from "../../types";
-let componentReferences: { [key: string]: Reference[] } = {};
+import {
+  getCssPropsTemplate,
+  getEventsTemplate,
+  getPartsTemplate,
+  getSlotsTemplate,
+} from "./description-templates.js";
+import { toKebabCase, removeQuoteWrappers } from "./utilities.js";
+import { getReferencesByComponent } from "./references.js";
+import type {
+  Attribute,
+  CssValue,
+  CustomElementsManifest,
+  Declaration,
+  Options,
+  TagAttribute,
+  Value,
+  VsCssProperty,
+} from "../../types";
 
-export function setComponentReferences(ts: any, node: any, moduleDoc: any) {
-  if (node.kind !== ts.SyntaxKind.ClassDeclaration) {
-    return;
+const EXCLUDED_TYPES = ["string", "boolean", "undefined", "number", "null"];
+const defaultLabels = {
+  slots: "Slots",
+  events: "Events",
+  cssProperties: "CSS Properties",
+  cssParts: "CSS Parts",
+  methods: "Methods",
+};
+
+export let config: Options = {
+  outdir: "./",
+  htmlFileName: "vscode.html-custom-data.json",
+  cssFileName: "vscode.css-custom-data.json",
+  exclude: [],
+  descriptionSrc: undefined,
+  slotDocs: true,
+  eventDocs: true,
+  cssPropertiesDocs: true,
+  cssPartsDocs: true,
+  methodDocs: true,
+  labels: {},
+  cssSets: [],
+};
+
+export function updateConfig(params: Options) {
+  config = { ...config, ...params };
+  config.labels = { ...defaultLabels, ...params?.labels };
+}
+
+export function getCssPropertyList(
+  customElementsManifest: CustomElementsManifest
+): VsCssProperty[] {
+  const components = getComponents(customElementsManifest);
+  return (
+    components?.map((component) => {
+      return (
+        component.cssProperties?.map((prop) => {
+          return {
+            name: prop.name,
+            description: prop.description,
+            values: getCssPropertyValues(prop?.type?.text),
+          };
+        }) || []
+      );
+    }) || []
+  ).flat();
+}
+
+export function getCssPartList(customElementsManifest: CustomElementsManifest) {
+  const components = getComponents(customElementsManifest);
+  return (
+    components?.map((component) => {
+      return (
+        component.cssParts?.map((prop) => {
+          return {
+            name: `::part(${prop.name})`,
+            description: prop.description,
+          };
+        }) || []
+      );
+    }) || []
+  ).flat();
+}
+
+export function getCssPropertyValues(value?: string): CssValue[] {
+  if (!value) {
+    return [];
   }
 
-  const references = getReferences(node);
-  updateReferences(references, node, moduleDoc);
+  if (value.trim().startsWith("set")) {
+    return getValueSet(value);
+  }
+
+  return getCssValues(value);
 }
 
-export function getReferencesByComponent(componentName: string) {
-  return componentReferences[componentName] ?? [];
+export function getValueSet(value: string): CssValue[] {
+  const setName = value.split(":")[1];
+  const valueSet =
+    config.cssSets?.find((x) => x.name.trim() === setName)?.values || [];
+
+  return valueSet.map((x) => {
+    if (typeof x === "string") {
+      return {
+        name: getCssNameValue(x),
+      };
+    } else {
+      x.name = getCssNameValue(x.name);
+      return x;
+    }
+  });
 }
 
-function getReferences(node: any) {
-  const docs = getDocsByTagName(node, "reference");
-  return docs
-    ?.map((tags: any) =>
-      tags?.map((doc: any) => {
-        const values = doc?.comment.split(/ - (.*)/s);
-
-        if (values && values.length > 1) {
-          return {
-            name: values[0].trim(),
-            url: values[1].trim(),
-          };
-        }
+export function getCssValues(value: string): CssValue[] {
+  return value
+    ? (value.includes("|") ? value.split("|") : value.split(",")).map((x) => {
+        const propName = x.trim();
+        return {
+          name: getCssNameValue(propName),
+        };
       })
+    : [];
+}
+
+function getCssNameValue(value: string) {
+  return !value ? "" : value.startsWith("--") ? `var(${value})` : value;
+}
+
+export function getTagList(customElementsManifest: CustomElementsManifest) {
+  const components = getComponents(customElementsManifest);
+  return components.map((component) => {
+    const slots = getSlotsTemplate(config, component.slots);
+    const events = getEventsTemplate(config, component.events);
+    const cssProps = getCssPropsTemplate(config, component.cssProperties);
+    const parts = getPartsTemplate(config, component.cssParts);
+
+    return {
+      name: component.tagName || toKebabCase(component.name),
+      description:
+        getDescription(component) + slots + events + cssProps + parts,
+      attributes: getComponentAttributes(component),
+      references: getReferencesByComponent(component.name),
+    };
+  });
+}
+
+function getDescription(component: Declaration) {
+  return (
+    (config.descriptionSrc
+      ? component[config.descriptionSrc]
+      : component.summary || component.description
+    )?.replaceAll("\\n", "\n") || ""
+  );
+}
+
+function getComponents(customElementsManifest: CustomElementsManifest) {
+  return customElementsManifest.modules
+    ?.map((mod) =>
+      mod?.declarations?.filter(
+        (dec: Declaration) =>
+          config.exclude &&
+          !config.exclude.includes(dec.name) &&
+          (dec.customElement || dec.tagName)
+      )
     )
     .flat();
 }
 
-function updateReferences(references: Reference[], node: any, moduleDoc: any) {
-  if (!references?.length) {
-    return;
-  }
+function getComponentAttributes(component: Declaration) {
+  const attributes: TagAttribute[] = [];
+  component?.attributes?.forEach((attr) => {
+    const existingAttr = attributes.find(
+      (x) => x.name === attr.name || x.name === attr.fieldName
+    );
+    if (existingAttr) {
+      return;
+    }
 
-  const className: string = node.name.getText();
-  const component: Declaration = moduleDoc?.declarations?.find(
-    (dec: Declaration) => dec.name === className
-  );
+    attributes.push({
+      name: attr.name || attr.fieldName,
+      description: attr.description,
+      values: getAttributeValues(attr),
+    } as TagAttribute);
+  });
 
-  componentReferences[component.name] = references as Reference[];
+  return attributes;
 }
 
-function getDocsByTagName(node: any, tagName: string) {
-  return node?.jsDoc?.map((doc: any) =>
-    doc?.tags?.filter((tag: any) => tag?.tagName?.getText() === tagName)
+function getAttributeValues(attr: Attribute): Value[] {
+  const value = attr.type?.text;
+  return !value
+    ? []
+    : (value.includes("|") ? value.split("|") : value.split(","))
+        .filter((type) => !EXCLUDED_TYPES.includes(type.trim()))
+        .map((type) => {
+          return {
+            name: removeQuoteWrappers(type),
+          } as Value;
+        });
+}
+
+function getMethods(component: Declaration) {
+  return component.members?.filter(
+    (member) =>
+      member.kind === "method" &&
+      member.privacy !== "private" &&
+      member.description?.length
   );
 }
